Restricte rollen som gir Prometheus tilgang til  scrape apper #84
https://github.com/nais/naiserator/issues/84

add Paths to AccessRule
make IstioPrometheusServiceAccount more specific
remove Prometheus rule from ServiceRoleBinding function
create ServiceRolePrometheus and ServiceRoleBindingPrometheus
test for the two functions

This makes it so that prometheus can't access the whole app. Now it's
only allowed to GET a specific metric path.

This is fixed by adding Paths to the AccessRule struct so we can
restrict paths that are allowed to be visited. Aswell as creating two
wrapper functions around ServiceRole and ServiceRoleBinding which
creates the more specicialized service roles and service role bindings
we need for prometheus.