Modified prometheus rolebinding spec, equality fix
Creating the prometheus rolebinding i initially overwrote the whole
spec, while really only the subjects need to be modified. Also, there
was a little pointless equality check which checked something that
always evaluates to true.

Before the creation of the servicerolebinding failed, because RoleRef
was missing, because I overwrite the whole spec and only specify
subject. Now that's fixed and creation of servicerole and
servicerolebinding works!