Restrict prometheus (#87)
* changed app.yaml to nais.yaml so it's current with the files in the examples/ dir

* add Paths to AccessRule
* make IstioPrometheusServiceAccount more specific
* remove Prometheus rule from ServiceRoleBinding function
* create ServiceRolePrometheus and ServiceRoleBindingPrometheus
* create CreateObjectMetaWithName
* test for the two functions

This makes it so that prometheus can't access the whole app. Now it's
only allowed to GET a specific metric path.

This is fixed by adding Paths to the AccessRule struct so we can
restrict paths that are allowed to be visited. Aswell as creating two
wrapper functions around ServiceRole and ServiceRoleBinding which
creates the more specicialized service roles and service role bindings
we need for prometheus.

ServiceRolePrometheus and ServiceRoleBindingPrometheus are standalone functions, which returns the almost the same as ServiceRole and SerivceRoleBinding at the time of writing. Only some of the parameters are modified to fit the specific prometheus rule.