[Bug Fix] Delete ingress objects
This bug is triggered when one or more ingresses are defined in the spec
and the application is updated to have 0 ingresses. Under these
conditions the previous ingresses are not deleted and the app keeps its
old ingresses.

https://github.com/nais/naiserator/issues/27

What I do is that I create a function cleanupFrutureRemovedResources,
which checks if the ingresses are to be removed during the app update.
If so, it proceeds to delete the old ingresses.