[Code Refactor] moved, renamed and simplified cleanupFutureRemovedResources
Ended up only needing to check the new app, so I moved the function
into synchronize(). Also renamed it to 'removeOrphanIngresses' as it's the
only thing it does now, but it could be extended in the future. The
function itself is also simpler; we only check the number of ingresses,
and based on that as well as if the delete operation was succesful we log and
return from the function.