Delete ingress object (#88)
* [Bug Fix] Delete ingress objects

This bug is triggered when one or more ingresses are defined in the spec
and the application is updated to have 0 ingresses. Under these
conditions the previous ingresses are not deleted and the app keeps its
old ingresses.

https://github.com/nais/naiserator/issues/27

What I do is that I create a function cleanupFrutureRemovedResources,
which checks if the ingresses are to be removed during the app update.
If so, it proceeds to delete the old ingresses.

* [Code Refactor] moved, renamed and simplified cleanupFutureRemovedResources

Ended up only needing to check the new app, so I moved the function
into synchronize(). Also renamed it to 'removeOrphanIngresses' as it's the
only thing it does now, but it could be extended in the future. The
function itself is also simpler; we only check the number of ingresses,
and based on that as well as if the delete operation was succesful we log and
return from the function.

* [Style] formatting, logging format directive and changing parameter order
Forgot to pass a parameter, changed it so it's passed and swaped the
order of the parameters at the same time. Also fixed some spacing in the
code.

* [Style] Remove a line, more informative logging inside removeOrphanIngresses

* fail fast if and change the log message format