[Feature] Added basic kafka and protobuf integrations This is by no means complete, but it's a start.
I've implemented a skeleton for the kafka configuration.
Added Makefiles and helpful developer docs in the README.
I've also started on the kafka package which shows the design
i'm thinking of going with, which is a producerLoop, listening on a
channel, and in some way receiving the protobuf Event struct on that
channel. The way it recieves the struct I've still not decided on.