[Refactor] Moved application Rollout check to naiserator package, etc.
Now the logic that checks if an app is deployed is moved to the
naiserator package, which is a much better fit. Now it's a continuous
check if the deployment has all of it's pods ready and when that
condition is met, a bool is sent over channel to signify that the
application is ready.

This is then again used by a function in the kafka package which waits
for 5 minutes for a bool to be sent over that channel before it times
out. If the bool is sent an event is then sent over to kafka.

Other than that I did som small changes to logging. But there's more
work to be done here.