[Refactor] Moved UUID generation into the syncronize loop
Now the uuid generation can fail and the sending of events would stop.

I also changed the NewDeploymentEvent function to use the app annotation
`nais.io/deployment-correlation-id` as it's correlationID. This way it's
set at the start and subsequent events all have that correlationID
automatically, as it's stored in the app.