[Feature] Added counter to events with same message
Fixes the repetition of events, seen in #100.

It reduces the spam by associating a counter with events that report the
same message. It works for all kinds of events so everything from
repeating errors to Successfull sync messages will be grouped together
and associated with a counter and a timespan in which all the errors
appeared.

What I've done:

I've expanded the reportEvent function to first get all the events for
the current app. Then match based on message and if it finds a matching
message it increments that events counter and updates its lastSeen
timestamp. Then it updates the event.

If it can't find the same event it creates a new one.

Since the feature is implemented in reportEvent it works for all kinds
of events not just Warnings/Errors

I've also made the descision to not to create a new event if the current
events for the app can't be fetched. This is to stop the creation of two
events with the same message and different counters.