Add counter to application events with same message (#102)
Fixes #100, the repetition of events in event log.

It reduces the spam by associating a counter with events that report the same message. It works for all kinds of events so everything from repeating errors to successful sync messages will be grouped together and associated with a counter and a timespan in which all the errors appeared.

Creation of a new event is aborted if the current events for the app can't be fetched.