[Feature] Added new clusterrole to restrict permissions for naiserator Fixes #17
I've created the naiserator-admin clusterrole, which get's it
permissions from a clusterroleselector. Following that I've created two
new clusterroles; one which is set to true by default, with permissions
to do everything naiserator has to do. And a conditional role that's
enabled based on if access policies are enabled for naiserator.

I've gotten the necessary permission from the api-server logs over what
naiserator has done over that last 7 months and looking through the
resorcecreator package.

It is also possible to simplify this drastically and create everything
under one clusterrole. As it's not that crucial whether or not
naiserator has access istio resouces if they don't exist.

Then again, one could argue that it is good to restrict naiserator
access to those resouces in clusters where they are available and
naiserator is not enabled for some reason.

All in all I think it's good to seperate them as it's structured better
that way and we have more fine graned control over the permissions,
which is what this whole issue is about.