feat: add option for loading default secrets for Vault
This commit adds an option to load default secrets for Vault
by specifying spec.vault.loadDefault (defaults to false to ensure
non-breaking behavior). Does not affect applications that do not specify
spec.vault.paths.

The motivation behind this is to simplify mounting of secrets for
developers. Currently, if spec.vault.paths is specified, default app
secrets are not mounted, meaning that the developers must manually
specify the kvPath and mountPath for their default secrets if they
need these in addition to other secrets they've specified.