Authoritative deletion (#163)
With this patch, Naiserator will assume ownership of all resources with the label `app=<application_name>`. All existing in-cluster resources that matches this label, but _is not present_ when generating resources for a new deployment, will be deleted. It only applies to resources where `ownerReferences` contains a reference to the owner application.

This merge only sports _logging_ of resources that would be deleted.