azureadapplication: generate hashed secret name with date suffix
This effectively restricts secret rotation to only occurs at most once
every 24 hours.

The rationale here is that secret rotation involves operations in
azurerator that may take 10's of seconds to perform and complete.
This incurs a cost in longer deployment times as pods will have to
wait for a new secret before starting. It is especially apparent in
development clusters where many single applications are deployed
frequently.