fix(google/sql): ensure that resource ID is set for additional users (#267)
* fix(google/sql): ensure that resource ID is set for additional users

As per https://cloud.google.com/config-connector/docs/reference/resource-docs/sql/sqluser,
the default behaviour for SQLUser is to use `metadata.name` as the name
of the SQL user. We've set this to be `$sqlinstance-$clean(user)-$hash` for
any additional users. However, this results in a mismatch between the
actual SQL username that is provisioned and the desired username
(the latter of which is also used in the generated secret) - and
ultimately causes authentication errors.

This commit fixes this by specifying the `resourceID` field and setting
the value of this to the desired username for additional users only.

* test: remove duplicate fields that are already evaluated