fix(poddisruptionbudget): use maxUnavailable instead of minAvailable (#268)
Previously, we set `minAvailable` in the PDB to be equal to the minimum desired
replica count - which in most cases results in the PDB not allowing any
disruptions because `count(pods) == minAvailable`. This effectively
halts draining of nodes containing pods affected by these PDBs, and in
the case of GKE will force-delete these pods when the grace period of 1
hour passes.

By setting the `maxUnavailable` to 1 we allow for a maximum of 1 disrupted
pod during upgrades. This comes with the caveat of effectively disabling PDBs
for applications that have set the minimum replica count to 1 and have not been
scaled up by the HPA at the time of the drain. However, one could argue that
users desiring high availability of their applications should specify at least
2 minimum replicas.