fix(poddisruptionbudget): skip creating PDB if min replicas is set to 1
Currently, if a PDB is created with min replica set to 1 and max > 1,
the PDB will disallow any disrupted pods if the deployment only has a
single active replica at the time of draining. This results in the
draining process of the node to halt, which is undesired. Applications
requiring zero disruptions should specify minimum 2 replicas.