wonderwall: use downward API to propagate pod IP
Use the pod IP instead of 127.0.0.1 when proxying to the application container.
The pod IP should be more consistent, as applications that should be exposed will either
bind to all interfaces or to just the interface matching the pod IP - refusing requests
via the loopback interface in the latter case.