Set logger for Manager
controller-runtime Managers now expect a Logger passed in that will be used for logging.
If not set, it will complain about no logger.
The passed in logger needs to be a logr.Logr instance, which sends logs to a LogSink.
We use Logrus2Logr, which implements the LogSink interface for logrus loggers.