charts: set texas.image from texas configmap or not at all
To get proper CD for texas, the image needs to come from the configmap defined by the texas build/deploy pipeline.
Naiserator should get the value from that configmap, or not at all.
The value should not be set in fasit at all.

This is the same reasoning and pattern as in eff9348f.