Avoid sync for apps that have not been synced with effective image yet
When this feature goes out, all apps will have an empty effective image, and would have been considered changed with the code as it was before.
This is unfortunate, as we only want to sync when there is an actual change, and in these cases we can wait for the hash to change before syncing and setting effective image.