Workload Image (#598)
* Add .dockerignore to reduce docker context

* Initial Tiltfile for local development

* Tiltfile: Auto-generate liberator charts on changes

* Tiltfile: Ensure naiserator stays in the context from tilt (re)load

* Get Image from external Image resource if not present in spec (WIP)

* Tiltfile: Hardcode single directory to avoid creating new tempdir all the time

* Tiltfile: Make naiserator startup optional

This allows for running naiserator in debugger instead of managed by tilt

* Add mise to project and set up tilt tasks

* Allow for long timeouts when running with tilt

This allows for slow stepping through debugger etc.

* Use image from Image resource if no image on Application/Naisjob

If there is no image on the resource when reconciling, look for a matching Image resource
and use the image from there.
The chosen image is set on Status.EffectiveImage when resource is saved to the cluster.

* Improve mise tilt tasks

* Add synchronizer test for applications with external image

* Use liberator from workload-image branch

* Latest liberator on main

* A bit too long for tests to run ...

* Avoid sync for apps that have not been synced with effective image yet

When this feature goes out, all apps will have an empty effective image, and would have been considered changed with the code as it was before.
This is unfortunate, as we only want to sync when there is an actual change, and in these cases we can wait for the hash to change before syncing and setting effective image.

* Give controller runtime a logger