feat(texas): reduce cpu requests to 10m
We generally see very infrequent peaks for CPU usage at around 20m.
The average load for applications with heavy use seem to hover
around 10-15m. Usage is even less (almost by an order of magnitude)
for the average application.

We reduce the requests here to reduce the impact on available CPUs
as we now inject Texas for all applications that use any authentication
provider. Most of these applications do not use Texas (yet).

As we don't set any limits, we should be able to utilize any unused CPU
that is available. Additionally, Texas will scale horizontally together
with the application.